Move Pattern tests to separate Pattern.cs file
Extracted Pattern-related tests from Rejigs.Tests.cs into a new Pattern.cs file for better organization and maintainability of the test suite.
